+++ CHANGELOG.rst
@@ -1,7 +1,7 @@
 History
 =======
 
-0.5.1 [Unreleased]
+0.5.1 (30-05-2024)
 ------------------
 
 * Added voice span support (#55)